+++ service.go
@@ -131,16 +131,16 @@ func sendHTMLFromProxySourceToQueue(rabbitConn *amqp.Connection) error {
 		if err != nil {
 			return err
 		}
-		htmlPayload := payload{
+		hPayload := htmlPayload{
 			HTML:    string(body),
 			FromURL: string(in),
 		}
-		htmlPayloadJSON, err := json.Marshal(htmlPayload)
+		hPayloadJSON, err := json.Marshal(hPayload)
 		if err != nil {
 			return err
 		}
 
-		if err := publish(chDest, queueProxySourceHTML, htmlPayloadJSON); err != nil {
+		if err := publish(chDest, queueProxySourceHTML, hPayloadJSON); err != nil {
 			return err
 		}
 		return nil
@@ -179,27 +179,27 @@ func processSourceHTML(rabbitConn *amqp.Connection, rdbForSites *redisDB, rdbFor
 			fmt.Println(time.Now().Format("2006-01-02T15:04:05"), "processSourceHTML end", time.Now().Sub(ts))
 		}(ts)
 
-		var htmlPayload payload
-		if err := json.Unmarshal(in, &htmlPayload); err != nil {
+		var hPayload htmlPayload
+		if err := json.Unmarshal(in, &hPayload); err != nil {
 			return err
 		}
 
-		proxies := findProxiesFromHTML(htmlPayload.HTML)
-		for _, proxy := range proxies {
-			changeType, err := rdbForProxies.set(proxy)
+		proxies := findProxiesFromHTML(hPayload.HTML)
+		for _, p := range proxies {
+			changeType, err := rdbForProxies.set(p)
 			if err != nil {
 				return err
 			}
 			switch changeType {
 			case redisChangeAdd, redisChangeUpdate:
-				if err := publish(chDestRawProxies, queueRawProxies, []byte(proxy)); err != nil {
+				if err := publish(chDestRawProxies, queueRawProxies, []byte(p)); err != nil {
 				}
 			}
 		}
 
-		urls := findURLsFromHTML(htmlPayload.HTML)
+		urls := findURLsFromHTML(hPayload.HTML)
 		for _, u := range urls {
-			if !urlsHaveSameDomain(htmlPayload.FromURL, u) {
+			if !urlsHaveSameDomain(hPayload.FromURL, u) {
 				continue
 			}
 			if !possibleForProxySourceURL(u) {
@@ -221,6 +221,57 @@ func processSourceHTML(rabbitConn *amqp.Connection, rdbForSites *redisDB, rdbFor
 	}
 
 	if err := consume(chSource, queueProxySourceHTML, handler); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func processRawProxy(rabbitConn *amqp.Connection) error {
+	chSource, err := rabbitConn.Channel()
+	if err != nil {
+		return err
+	}
+	defer chSource.Close()
+
+	chDest, err := rabbitConn.Channel()
+	if err != nil {
+		return err
+	}
+	defer chDest.Close()
+
+	handler := func(in []byte) error {
+		fmt.Println(time.Now().Format("2006-01-02T15:04:05"), "processRawProxy - start")
+		ts := time.Now()
+		defer func(ts time.Time) {
+			fmt.Println(time.Now().Format("2006-01-02T15:04:05"), "processRawProxy end", time.Now().Sub(ts))
+		}(ts)
+
+		pType, anonymous, err := checkProxy(string(in))
+		if err != nil {
+			return err
+		}
+		tsNow := time.Now()
+		tsNowRFC3339 := tsNow.Format(time.RFC3339)
+
+		p := proxyPayload{
+			URL:       string(in),
+			Anonymous: anonymous,
+			Type:      pType.verbose(),
+			TS:        tsNowRFC3339,
+		}
+		payloadJSON, err := json.Marshal(p)
+		if err != nil {
+			return err
+		}
+
+		if err := publish(chDest, queueProcessedProxies, payloadJSON); err != nil {
+		}
+
+		return nil
+	}
+
+	if err := consume(chSource, queueRawProxies, handler); err != nil {
 		return err
 	}