+++ util.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"net/url"
-	"path/filepath"
 )
 
 func uniqueArrayOfValues(arr []string) (unique []string) {
@@ -23,31 +22,6 @@ func isExist(val string, arr []string) bool {
 		}
 	}
 	return false
-}
-
-func possibleForProxySourceURL(u string) bool {
-	uParsed, err := url.Parse(u)
-	if err != nil {
-		return false
-	}
-
-	forbiddenDomains := []string{
-		"dictionary.cambridge.org", "cambridge.org", "www.cambridge.org",
-		"microsoft.com", "www.microsoft.com",
-		"theunfolder.com", "www.theunfolder.com",
-		"zen.yandex.ru"}
-	if isExist(uParsed.Hostname(), forbiddenDomains) {
-		return false
-	}
-
-	uParsed.RawQuery = ""
-	uParsed.RawFragment = ""
-	forbiddenExts := []string{".css", ".ico", ".png", ".jpg", ".jpeg", ".webp", ".gif", ".js", ".ts", ".woff", ".woff2"}
-	if isExist(filepath.Ext(uParsed.String()), forbiddenExts) {
-		return false
-	}
-
-	return true
 }
 
 func urlsHaveSameDomain(u1, u2 string) bool {